Finally got the attack and release functionality working. I had to increase the clock speed to 16MHz and halve the sample interrupt frequency to make it work. I suspect the interrupt took longer than the time between interrupts so it got caught up in an infinite loop and stopped outputting samples.

+++ wave_gen.h
@@ -1,7 +1,6 @@
 #include <avr/io.h>
-#include <avr/pgmspace.h>
 
 uint8_t getMaxVoices();
 void setJump(uint8_t index, int jumpSize);
-void setAmplitude(uint8_t index, uint8_t ampl);
+void setAmplitude(uint8_t index, uint16_t ampl);
 void setWaveOutput(uint8_t waveSetting);
